refactor: ReadNews 도메인 분리 api Test에 적용
